refactor duplicate nodes and edges #286
