+++ band/graph/graph.h
@@ -1,25 +1,13 @@
-#ifndef BAND_MODEL_GRAPH_GRAPH_H_
-#define BAND_MODEL_GRAPH_GRAPH_H_
+#ifndef BAND_GRAPH_GRAPH_H_
+#define BAND_GRAPH_GRAPH_H_
 
-#include "absl/status/status.h"
-#include "absl/status/statusor.h"
+#include "band/graph/graph_interface.h"
 #include "band/graph/node.h"
 #include "band/model.h"
+#include "absl/status/status.h"
+#include "absl/status/statusor.h"
 
 namespace band {
-
-using Edge = std::pair<size_t, size_t>;
-
-class IGraph {
- public:
-  virtual std::vector<std::shared_ptr<Node>> nodes() const = 0;
-  virtual std::vector<Edge> edges() const = 0;
-  virtual std::vector<size_t> GetParents(size_t node_id) const;
-  virtual std::vector<size_t> GetChildren(size_t node_id) const;
-  virtual std::shared_ptr<Node> GetNodeById(size_t id) const {
-    return nodes()[id];
-  };
-};
 
 class Graph : public IGraph {
   friend class GraphBuilder;
@@ -27,24 +15,19 @@ class Graph : public IGraph {
  public:
   std::string GetGraphVizText() const;
   absl::Status SaveGraphViz(std::string path) const;
-
-  std::vector<std::shared_ptr<Node>> nodes() const override { return nodes_; }
-  std::vector<Edge> edges() const override { return edges_; }
-
   std::vector<size_t> GetTopologicalOrder() const;
 
  private:
   Graph(std::string name, std::vector<std::shared_ptr<Node>> nodes,
         std::vector<Edge> edges)
-      : name_(name), edges_(edges) {
+      : IGraph(name) {
     for (auto& node : nodes) {
       nodes_.push_back(node);
     }
+    for (auto& edge : edges) {
+      edges_.push_back(edge);
+    }
   }
-
-  std::string name_;
-  std::vector<std::shared_ptr<Node>> nodes_;
-  std::vector<Edge> edges_;
 };
 
 }  // namespace band
